+++ alacritty/src/cli.rs
@@ -298,8 +298,7 @@ mod tests {
     fn dynamic_title_overridden_by_options() {
         let mut config = Config::default();
 
-        let mut options = Options::default();
-        options.title = Some("foo".to_owned());
+        let options = Options { title: Some("foo".to_owned()), ..Options::default() };
         options.override_config(&mut config);
 
         assert!(!config.ui_config.window.dynamic_title);